Normalizing query names on cache lookup
Until now, the lookup was performed in the client-go's cache
using query name as it was received. This worked fine for all
normal queries but broked when queries contained uppercase letters.

One notable example of what got broked was Let's Encrypt. Their
DNS resolver would lookup the target domain name using a mix of
lower and uppwer case letters, which would result in a cache miss
and would fail with NXDOMAIN for the target domain.

This change converts all cache lookup keys to lower case.